requirements.txt
tqdm
pyserial
numpy
matplotlib
